Updates
Copied CSS folder to root and fixed the link to it on pages.
Fixed the link to the logo.

Added a new CSS file, Races.css, directed Terran, protoss, zerg html pages to that.

Added the base layout to Terran, Protoss and Zerg HTML pages.

Removed stuff from the main page that was leftover from the layout.

Added random input fields to register and login pages.

Added unit names to plans.